requirements.txt
aiofiles==23.1.0
aiogram==3.1.1
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.5
defusedxml==0.8.0rc2
Django==4.0
django-cors-headers==4.3.1
django-extensions==3.2.3
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
drf-yasg==1.21.7
flake8==6.1.0
flake8-broken-line==1.0.0
flake8-isort==6.1.1
frozenlist==1.4.0
graphviz==0.20.1
idna==3.4
inflection==0.5.1
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.2
magic-filter==1.0.12
MarkupSafe==2.1.3
mccabe==0.7.0
multidict==6.0.4
mypy==1.6.1
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.2
pep8-naming==0.13.3
Pillow==10.1.0
prettytable==3.9.0
pycodestyle==2.11.1
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
pydot==1.4.2
pyflakes==3.1.0
PyJWT==2.8.0
pyparsing==3.1.1
python-dotenv==1.0.0
python3-openid==3.2.0
pytils==0.4.1
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.5.0
sorl-thumbnail==12.10.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
wcwidth==0.2.9
yarl==1.9.2
